sql: move rand datum functions to testutils.go
